Use correct import paths for QML engine for palettes
