+++ mscore/CMakeLists.txt
@@ -489,6 +489,7 @@ add_executable ( ${ExecutableName}
       extension.cpp extension.h
       tourhandler.cpp
       script/script.cpp script/scriptentry.cpp script/testscript.cpp script/recorderwidget.cpp
+      qml/msqmlengine.cpp
       qmldockwidget.cpp
 
       ${WIDGETS_SOURCE_FILES}
